feat(msg): only use clientLocalTimestamp as sequenceNumber to assign replies to request

+++ src/main/java/isos/communication/client/RequestReplyHandler.java
@@ -21,7 +21,7 @@ public interface RequestReplyHandler<T> {
    */
   ClientMessageWrapper createRequest(byte[] payload);
 
-  int getSequenceId();
+  long getSequenceId();
 
   void waitForResponse() throws InterruptedException;
